refactor expanded view to functional, added pointer when hovering the expanded view buttons, installed react-prismazoom
