+++ package.json
@@ -25,6 +25,7 @@
     "eslint-plugin-react": "^7.22.0",
     "eslint-plugin-react-hooks": "^4.2.0",
     "jest": "^26.6.3",
+    "react-prismazoom": "^2.0.1-rc0",
     "react-social-icons": "^4.1.0",
     "react-test-renderer": "^15.7.0",
     "webpack": "^2.2.1"
